+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vformore/core
 
+## 1.0.4
+
+### Patch Changes
+
+- fix types for createInstance
+
 ## 1.0.3
 
 ### Patch Changes